code/llm/requirements.txt
pandas
scikit-lear
numpy
faiss-cpu
sentence-transformers
pypdf
openai
tqdm
requests
yaml